22/11/2008 (1.8.1.183) ---------- - Assets: Plug-in will no longer crash if it can't find a report file - Map: Fixed error caused by setting a WP when no start and end system has been selected - Map: Fixed docking of Exclusion controls - Map: Fixed docking of Waypoint controls - Assets: Fixed sort order of Quantity column - Core: Fixed multiple colour selection settings from not responding when double clicking - HQF: Now handled errors that could occur when exporting formats to the clipboard - HQF: Added "Flush" command to the filestream writing the fittings, profiles and HQF pilot data

+++ EveHQ.Assets/frmAssets.Designer.vb
@@ -257,7 +257,7 @@ Partial Class frmAssets
         Me.ContainerListViewColumnHeader2.ContentAlign = System.Drawing.ContentAlignment.MiddleRight
         Me.ContainerListViewColumnHeader2.CustomSortTag = Nothing
         Me.ContainerListViewColumnHeader2.DisplayIndex = 5
-        Me.ContainerListViewColumnHeader2.SortDataType = DotNetLib.Windows.Forms.SortDataType.[Integer]
+        Me.ContainerListViewColumnHeader2.SortDataType = DotNetLib.Windows.Forms.SortDataType.[Double]
         Me.ContainerListViewColumnHeader2.Tag = Nothing
         Me.ContainerListViewColumnHeader2.Text = "Quantity"
         Me.ContainerListViewColumnHeader2.Width = 100
